feat(email): collocate Booking Reminder variables with template (#51)
Move the Swedish-locale variable builder onto EMAIL_TEMPLATES.bookingReminder.build
so template metadata and its variable construction live as one deep unit. Add a
typed EmailTemplate<Input> wrapper around EmailTemplateSpec, with trivial { URL }
pass-throughs on verifyEmail and resetPassword. Reminder scheduler stops
importing the standalone builder; the reminder.email module is gone. Migrated
the six Swedish-locale tests to a new email.templates.spec.ts collocated with
the registry.
